--- src/shared-types.ts
@@ -1,30 +1,0 @@
-export type LinkKeys =
-  | 'signin'
-  | 'signup'
-  | 'uiTest'
-  | 'visualTest'
-  | 'interactionTest'
-  | 'storybook'
-  | 'playwright'
-  | 'cypress'
-  | 'turboSnap'
-  | 'uiReview'
-  | 'publish'
-  | 'figmaPlugin'
-  | 'frontendTeams'
-  | 'designSystems'
-  | 'digitalAgencies'
-  | 'aboutChromatic'
-  | 'careers'
-  | 'security'
-  | 'enterprise'
-  | 'netlify'
-  | 'monday'
-  | 'collective'
-  | 'ezcater'
-  | 'blog'
-  | 'changelog'
-  | 'frontendTestingGuide'
-  | 'docs'
-  | 'pricing'
-  | 'sales';
